requirements.txt
mkdocs-material
mkdocs-graphviz
mkdocs-static-i18n
